Correct sorting behavior of Sources Tab.
