+++ src/main/java/decodes/dbeditor/SourceEditPanel.java
@@ -1,3 +1,18 @@
+/**
+ * Copyright 2024 The OpenDCS Consortium and contributors
+ *
+ *  Licensed under the Apache License, Version 2.0 (the "License");
+ *  you may not use this file except in compliance with the License.
+ *  You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *  Unless required by applicable law or agreed to in writing, software
+ *  distributed under the License is distributed on an "AS IS" BASIS,
+ *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *  See the License for the specific language governing permissions and
+ *  limitations under the License.
+ */
 package decodes.dbeditor;
 
 import java.awt.*;
@@ -368,11 +383,10 @@ public class SourceEditPanel extends DbEditorTab
 		// This also updates the SourceListPanel.
 		getDb().dataSourceList.getList().remove(origDataSource);
 		getDb().dataSourceList.add(dataSource);
-		parent.getSourcesListPanel().resort();
+		parent.getSourcesListPanel().tableModel.fireTableDataChanged();
 
 		// Replace DataSource in every Group using this data source
-		for(Iterator<DataSource> it = getDb().dataSourceList.getList().iterator();
-			it.hasNext(); )
+		for(Iterator<DataSource> it = getDb().dataSourceList.getList().iterator(); it.hasNext(); )
 		{
 			DataSource ds = it.next();
 			if (ds.getName().equalsIgnoreCase(dataSource.getName()))
