+++ src/main/java/decodes/dbeditor/SourcesListPanel.java
@@ -1,45 +1,38 @@
-/*
-*  $Id$
-*  
-*  Open Source Software
-*  
-*  $Log$
-*  Revision 1.1.1.1  2014/05/19 15:28:59  mmaloney
-*  OPENDCS 6.0 Initial Checkin
-*
-*  Revision 1.5  2013/03/21 18:27:40  mmaloney
-*  DbKey Implementation
-*
-*/
+/**
+ * Copyright 2024 The OpenDCS Consortium and contributors
+ *
+ *  Licensed under the Apache License, Version 2.0 (the "License");
+ *  you may not use this file except in compliance with the License.
+ *  You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *  Unless required by applicable law or agreed to in writing, software
+ *  distributed under the License is distributed on an "AS IS" BASIS,
+ *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *  See the License for the specific language governing permissions and
+ *  limitations under the License.
+ */
 package decodes.dbeditor;
 
 import java.awt.*;
 import java.awt.event.*;
 import javax.swing.*;
-import javax.swing.table.*;
 
-import ilex.util.PropertiesUtil;
 import decodes.gui.*;
-import java.util.Comparator;
-import java.util.Collections;
 import java.util.Iterator;
 import java.util.ResourceBundle;
 
 import ilex.util.LoadResourceBundle;
 
-import decodes.db.DatabaseObject;
-import decodes.db.DatabaseException;
 import decodes.db.Database;
 import decodes.db.DataSource;
-import decodes.db.DataSourceList;
-import decodes.db.RoutingSpec;
-import decodes.db.Constants;
+import decodes.dbeditor.sources.DataSourceListTableModel;
 
 /**
 DBEDIT panel containing sorting list of data sources.
 */
-public class SourcesListPanel extends JPanel
-	implements ListOpsController
+public class SourcesListPanel extends JPanel implements ListOpsController
 {
 	static ResourceBundle genericLabels = DbEditorFrame.getGenericLabels();
 	static ResourceBundle dbeditLabels = DbEditorFrame.getDbeditLabels();
@@ -57,9 +50,10 @@ public class SourcesListPanel extends JPanel
 	{
 		try
 		{
-			tableModel = new DataSourceListTableModel(this);
-			dataSourceListTable = new SortingListTable(tableModel,
-				new int[] { 25, 15, 50, 10 });
+			tableModel = new DataSourceListTableModel(this,Database.getDb());
+			dataSourceListTable = new JTable(tableModel);
+			dataSourceListTable.setAutoCreateRowSorter(true);
+			//new int[] { 25, 15, 50, 10 });
 	 		dataSourceListTable.getSelectionModel().setSelectionMode(
 		    	ListSelectionModel.SINGLE_SELECTION);
 
@@ -108,8 +102,11 @@ public class SourcesListPanel extends JPanel
 	{
 		int idx = dataSourceListTable.getSelectedRow();
 		if (idx == -1)
+		{
 			return null;
-	    return tableModel.getObjectAt(idx);
+		}
+		int modelRow = dataSourceListTable.convertRowIndexToModel(idx);
+	    return tableModel.getObjectAt(modelRow);
 	}
 
 	/** Called when the 'Open' button is pressed. */
@@ -208,10 +205,10 @@ public class SourcesListPanel extends JPanel
 
 		// Also don't allow delete if this source is a group member of 
 		// another data source.
-		for(Iterator it = Database.getDb().dataSourceList.iterator();
+		for(Iterator<DataSource> it = Database.getDb().dataSourceList.iterator();
 			it.hasNext(); )
 		{
-			DataSource x = (DataSource)it.next();
+			DataSource x = it.next();
 			if (x == ds)
 				continue;
 			if (x.isInGroup(ds.getName()))
@@ -267,168 +264,5 @@ public class SourcesListPanel extends JPanel
 			dbtp.add(title, newTab);
 			dbtp.setSelectedComponent(newTab);
 		}
-	}
-
-	/** Resorts the list by the specified column. */
-	public void resort()
-	{
-		this.tableModel.resort();
-	}
-}
-
-
-class DataSourceListTableModel extends AbstractTableModel
-	implements SortingListTableModel
-{
-	private String colNames[] = 
-	{
-		SourcesListPanel.genericLabels.getString("name"),
-		SourcesListPanel.genericLabels.getString("type"),
-		SourcesListPanel.dbeditLabels.getString("SourcesListPanel.args"),
-		SourcesListPanel.dbeditLabels.getString("SourcesListPanel.usedBy")
-	};
-	private SourcesListPanel panel;
-	private int lastSortColumn = -1;
-	private DataSourceList theList;
-
-	public DataSourceListTableModel(SourcesListPanel slp)
-	{
-		super();
-		this.panel = slp;
-		theList = Database.getDb().dataSourceList;
-		theList.countUsedBy();
-		this.sortByColumn(0);
-		refill();
-	}
-
-	void refill()
-	{
-	}
-
-	public int getRowCount()
-	{
-		return theList.size();
-	}
-
-	DataSource getObjectAt(int r)
-	{
-		return (DataSource)getRowObject(r);
-	}
-
-	public Object getRowObject(int r)
-	{
-		if (r >= 0 && r < getRowCount())
-			return theList.getList().get(r);
-		else return null;
-	}
-
-	void addDataSource(DataSource ds)
-	{
-		theList.add(ds);
-		fireTableDataChanged();
-	}
-
-	void deleteObject(DataSource ds)
-	{
-		theList.remove(ds);
-		try { Database.getDb().getDbIo().deleteDataSource(ds); }
-		catch(DatabaseException e)
-		{
-			if ( ds.getId() != Constants.undefinedId ) 
-				TopFrame.instance().showError(e.toString());
-		}
-		fireTableDataChanged();
-	}
-
-	public int getColumnCount()
-	{
-		int r = colNames.length;
-		return r;
-	}
-
-	public String getColumnName(int c)
-	{
-		return colNames[c];
-	}
-
-	public boolean isCellEditable(int r, int c) { return false; }
-
-	public Object getValueAt(int r, int c)
-	{
-		DataSource ds = getObjectAt(r);
-		if (ds == null)
-			return "";
-		else
-			return getDsColumn(ds, c);
-	}
-
-	public static String getDsColumn(DataSource ds, int c)
-	{
-		switch(c)
-		{
-		case 0: return ds.getName();
-		case 1:
-			return ds.dataSourceType == null ? "" : ds.dataSourceType;
-		case 2:
-			return ds.getDataSourceArgDisplay() == null ? "" : ds.getDataSourceArgDisplay();
-		case 3:
-			return "" + ds.numUsedBy;
-		default: return "";
-		}
-	}
-
-	public void resort()
-	{
-		if (lastSortColumn != -1)
-			sortByColumn(lastSortColumn);
-	}
-
-	public void sortByColumn(int c)
-	{
-		lastSortColumn = c;
-		Collections.sort(theList.getList(), new DataSourceComparator(c));
-		fireTableDataChanged();
-	}
-
-	void replace(DataSource oldOb, DataSource newOb)
-	{
-		theList.remove(oldOb);
-		theList.add(newOb);
-		if (lastSortColumn != -1)
-			sortByColumn(lastSortColumn);
-		else
-			fireTableDataChanged();
-	}
-
-}
-
-class DataSourceComparator implements Comparator
-{
-	int column;
-
-	public DataSourceComparator(int column)
-	{
-		this.column = column;
-	}
-
-	/**
-	 * Compare the eqMod names of the specified type.
-	 */
-	public int compare(Object ob1, Object ob2)
-	{
-		if (ob1 == ob2)
-			return 0;
-		DataSource ds1 = (DataSource)ob1;
-		DataSource ds2 = (DataSource)ob2;
-
-		String s1 = DataSourceListTableModel.getDsColumn(ds1, column);
-		String s2 = DataSourceListTableModel.getDsColumn(ds2, column);
-
-		return s1.compareToIgnoreCase(s2);
-	}
-
-	public boolean equals(Object ob)
-	{
-		return false;
-	}
+	}	
 }
